+++ src/django_sysinfo/api.py
@@ -16,7 +16,7 @@ from django.conf import settings
 from django.db import connections
 from django.utils.module_loading import import_string
 
-from django_sysinfo.compat import get_istalled_apps
+from django_sysinfo.compat import get_installed_apps
 from django_sysinfo.utils import get_network, humanize_bytes
 
 from .conf import config
@@ -68,13 +68,17 @@ def _get_database_infos(conn):
 def get_databases(**kwargs):
     databases = OrderedDict()
     for alias in connections:
-        conn = connections[alias]
         db = OrderedDict()
-        db["engine"] = conn.settings_dict.get("ENGINE")
-        db["host"] = "%(HOST)s:%(PORT)s" % conn.settings_dict
-        db["name"] = conn.settings_dict.get("NAME")
-        db.update(_get_database_infos(conn))
-        databases[alias] = db
+        try:
+            conn = connections[alias]
+            db["engine"] = conn.settings_dict.get("ENGINE")
+            db["host"] = "%(HOST)s:%(PORT)s" % conn.settings_dict
+            db["name"] = conn.settings_dict.get("NAME")
+            db.update(_get_database_infos(conn))
+        except Exception as e:
+            db["error"] = str(e)
+        finally:
+            databases[alias] = db
     return databases
 
 
@@ -110,6 +114,27 @@ def get_python(**kwargs):
     p["info"] = sys.version
     p["maxunicode"] = (sys.maxunicode,
                        {True: "OK", False: "WARN"}[sys.maxunicode > 0xffff])
+    return p
+
+
+def get_mail(**kwargs):
+    def check():
+        from django.core.mail import get_connection
+        try:
+            conn = get_connection(fail_silently=False)
+            conn.open()
+            ret = "OK"
+            conn.close()
+        except Exception as e:
+            ret = str(e)
+        return ret
+
+    p = OrderedDict()
+    p["backend"] = settings.EMAIL_BACKEND
+    p["host"] = "{0}:{1}".format(settings.EMAIL_HOST, settings.EMAIL_PORT)
+    p["tls"] = getattr(settings, "USE_TLS", False)
+    p["ssl"] = getattr(settings, "USE_SSL", False)
+    p["status"] = check()
     return p
 
 
@@ -150,7 +175,9 @@ def get_project(**kwargs):
         project["CACHES"] = get_caches_info()
 
     if config.installed_apps:
-        project["installed_apps"] = get_istalled_apps()
+        project["installed_apps"] = get_installed_apps()
+    if config.mail:
+        project["mail"] = get_mail(**kwargs)
     return project
 
 
@@ -159,12 +186,65 @@ def get_os(**kwargs):
             "name": os.name}
 
 
+def run_check(id, request=None, fail_silently=True, fail_status=500):
+    status = 200
+    try:
+        v = config.checks[id]
+        if isinstance(v, six.string_types):
+            c = import_string(v)
+            ret, status = c(request)
+        elif callable(v):
+            ret, status = v(request)
+        else:
+            ret = v
+    except Exception as e:
+        ret = "ERROR"
+        status = fail_status
+        logger.exception(e)
+        if settings.DEBUG:
+            ret = str(e)
+        if not fail_silently:
+            raise
+
+    return ret, status
+
+
+def get_checks(request=None):
+    checks = {}
+    if config.checks:
+        for k, v in config.checks.items():
+            checks[k] = run_check(k)
+
+    return checks
+
+
+def get_extra(config, request=None):
+    extras = {}
+    for k, v in config.extra.items():
+        try:
+            if isinstance(v, six.string_types):
+                c = import_string(v)
+                extras[k] = c(request)
+            elif callable(v):
+                extras[k] = v(request)
+            else:
+                extras[k] = v
+        except Exception as e:
+            logger.exception(e)
+            if settings.DEBUG:
+                extras[k] = str(e)
+    return extras
+
+
 handlers = OrderedDict([("host", get_host),
                         ("os", get_os),
                         ("python", get_python),
                         ("modules", get_modules),
                         ("project", get_project),
-                        ("databases", get_databases)])
+                        # ("databases", get_databases),
+                        ("extra", get_extra),
+                        ("checks", get_checks)])
+
 valid_sections = handlers.keys()
 
 
@@ -178,25 +258,7 @@ def get_sysinfo(request):
 
     for section in sections:
         if section in valid_sections and getattr(config, section):
-            data[section] = handlers[section](config=config)
-
-    if config.extra:
-        extras = {}
-        for k, v in config.extra.items():
-            try:
-                if isinstance(v, six.string_types):
-                    c = import_string(v)
-                    extras[k] = c(request)
-                elif callable(v):
-                    extras[k] = v(request)
-                else:
-                    extras[k] = v
-            except Exception as e:
-                logger.exception(e)
-                if settings.DEBUG:
-                    extras[k] = str(e)
-
-        data["extra"] = extras
+            data[section] = handlers[section](config=config, request=request)
 
     return data