+++ src/django_sysinfo/compat.py
@@ -9,37 +9,10 @@ import django
 
 logger = logging.getLogger(__name__)
 
-if django.VERSION[:2] == (1, 6):
-    def get_istalled_apps():
-        installed_apps = []
-        from django.conf import settings
-
-        for app_name in settings.INSTALLED_APPS:
-            try:
-                mod = __import__(app_name)
-                installed_apps.append([app_name,
-                                       get_distribution(mod.__name__).version])
-            except:
-                pass
-        return installed_apps
-
-    from django.http import HttpResponse
-    from django.core.serializers.json import DjangoJSONEncoder
-    import json
-
-
-    class JsonResponse(HttpResponse):
-        def __init__(self, data, encoder=DjangoJSONEncoder, safe=True, **kwargs):
-            kwargs.setdefault('content_type', 'application/json')
-            data = json.dumps(data, cls=encoder)
-            super(JsonResponse, self).__init__(content=data, **kwargs)
-
-elif django.VERSION[1] in [7, 8, 9, 10]:
+if django.VERSION[1] in [7, 8, 9, 10, 11]:
     from django.apps import apps
-    from django.http import JsonResponse, HttpResponse
 
-
-    def get_istalled_apps():
+    def get_installed_apps():
         installed_apps = []
         for app_config in apps.get_app_configs():
             try:
@@ -49,5 +22,5 @@ elif django.VERSION[1] in [7, 8, 9, 10]:
                 pass
         return installed_apps
 
-else:
+else:  # pragma: no cover
     raise EnvironmentError('Django version not supported')