+++ tests/demo/demoproject/models.py
@@ -4,3 +4,7 @@ from __future__ import absolute_import, unicode_literals
 
 def test_sysinfo(request):
     return 123
+
+
+def test_check(request):
+    return True, 200